requirements.txt
python-decouple
requests
BeautifulSoup4
html5lib